requirements.txt
streamlit
google-generativeai
langchain-community
langchain-google-genai
pypdf
faiss-cpu
openai
streamlit-TTS